193 Updating the Route(Section 13: Navigation From Scratch)
All right, so now the last thing we have to do is make sure that any time on location changed function

is called.

We have to tell somehow the root component to update itself or to really just re render itself when

it re renders itself.

We're going to fetch whatever the current path name is and use that to decide whether or not this root

component should now display its child or hide its.

Now, the Khoder we're about to write inside of here is going to be just a small bit confusing.

And the reason for that is that we're going to create a piece of state that essentially follows or tracks

whatever the value of window location Patnaik is.

This property right here is always kept perfectly up to date.

In other words, whenever you look at window, location, path, name, it's always gonna reflect exactly

whatever the current path name is inside the address bar.

The reason I'm saying that this is going to be a little bit confusing is that we're essentially introducing

a piece of state that is just has the sole purpose of trying to get the root component to re render

itself as the only job for this piece of state, because we could otherwise just refer to window location,

path name to figure out what the current path name is.

Now, if all this doesn't make any sense.

Don't sweat it.

Let's just add in some code right away.

All right.

So we need to add in a new piece of state, which means we're going to import the new state helper at

the top.

All then declare a piece of state called current path and set for in path.

And I'm going to use a default value for this of window, location.

Name.

Again, don't forget, that name has a lowercase N.

Then whenever our location changes, I can remove that council log and we will call set current path

and update its value to be whatever the current path name is inside of that address bar.

But just a moment ago, I said, yeah, this piece of state is gonna be all a bit confusing because

it's really just tracking weather path name is.

And you could see that very easily right away whenever this root component component is first displayed

on the screen.

Current path is set to be whatever path name is.

And then whenever the pattern changes, well, we just update current path to be once again, whatever

path name is.

Well, the only reason this piece of state exists is just to get our root to update.

That's it.

Now, finally, down here at the return statement, we could update this to be Brent path or technically

it could also stay as window location, path name as well.

It really doesn't make a difference, but I'm going to change it to be current path anyways, just to

make the component make a little bit more sense.

OK, so let's save this now and do a quick test.

The back inside my browser.

I'm going to click on Translate, and as soon as I do notice how the URL changes and the content on

the screen changes as well.

I can click on dropdown.

I can click on accordion and search.

And as I click around, not only does the URL change, but I also get the change in the content on the

screen as well.

Now, as you click around, if you go to that search component, you might see a warning like disappear.

If you do, don't sweat it.

It's because he essentially are your request or search that Wikipedia API is being resolved only after

we have navigated away.

You'll probably only see that warning if you're clicking between these different tabs very quickly.

This is a warning that is really, really challenging to fix.

So right now, we're just not going to worry about it too much.

OK.

So this is pretty much it.

We've now got some working navigation.

Now, there's one thing I want to take care of.

Whenever a user sees a link like this right here, some users.

I say some.

Who knows, maybe you do this.

Maybe you don't.

I personally use this feature a ton in your browser.

Normally, whenever you find a link, you should be able to hold down command on your keyboard if you're

on Mac OS and click it or hold down control if you're on windows and click it.

And in either case, that link should be opened up in a new tab, pretty much like if I.

Right.

Click on this thing and then hit the open link in new tab.

That is some expected behavior.

But right now, if I hold down command and click on that link, that definitely does not happen.

I think we still need to do a tiny little bit of work just to restore that feature.

And then after that, I think that this application is pretty much set.

All right.

Let's take a quick pause right here and fix that up in just a moment.